fix non-critical exception and add HikariCP configuration

+++ src/main/java/by/grodmir/IT_project_hub/infrastructure/repository/RefreshTokenJpaRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface RefreshTokenJpaRepository extends JpaRepository<RefreshTokenJpaEntity, Integer> {
+public interface RefreshTokenJpaRepository extends JpaRepository<RefreshTokenJpaEntity, Long> {
     Optional<RefreshTokenJpaEntity> findByToken(String token);
     void deleteAllByUsername(String username);
 }
